requirements.txt
# Paquetes principales
streamlit==1.24.0
pandas>=1.5.0
plotly>=5.11.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.2.0

# AWS/Backblaze B2
boto3>=1.26.0
botocore>=1.29.0

# Procesamiento de datos
openpyxl>=3.1.0  # Para soporte de Excel
python-dotenv>=1.0.0  # Para manejo de variables de entorno
xlrd>=2.0.1  # Para leer archivos Excel antiguos

# Utilidades
python-dateutil>=2.8.2
pytz>=2022.7

# Desarrollo
typing-extensions>=4.4.0

# Seguridad
urllib3>=1.26.0
requests>=2.28.0